app js modified
