--- src/app.js
@@ -1,11 +1,0 @@
-import express from "express";
-import userRouter from "./routes/user.js";
-
-import "./database.js";
-
-const app = express();
-app.use(express.json());
-
-app.use("/user", userRouter);
-
-export default app;
